refactor(tui-render): split app into screen modules
