+++ tui/crates/tui-kit-render/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
 name = "tui-kit-render"
 version = "0.1.0"
-edition = "2021"
-rust-version = "1.75"
+edition = "2024"
+rust-version = "1.85"
 license = "MIT"
 description = "Ratatui-based UI layer for tui-kit"
 
@@ -13,6 +13,7 @@ path = "src/lib.rs"
 [dependencies]
 anyhow = "1"
 ratatui = "0.29"
+unicode-width = "0.2"
 crossterm = "0.28"
 fuzzy-matcher = "0.3"
 tui-kit-model = { path = "../tui-kit-model" }
