+++ tui/crates/tui-kit-render/src/ui/app/shared/mod.rs
@@ -1,0 +1,5 @@
+//! Shared app-level rendering helpers reused across screens.
+
+pub mod header;
+pub mod layout;
+pub mod status;
