+++ tui/crates/tui-kit-render/src/ui/app/types.rs
@@ -70,8 +70,6 @@ pub struct CreateOverlayText {
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct SettingsOverlayText {
     pub title: String,
-    pub theme_label: String,
-    pub theme_action_key: String,
     pub close_hint: String,
 }
 
@@ -89,8 +87,6 @@ pub struct StatusText {
     pub commands_label: String,
     pub tabs_label: String,
     pub quit_label: String,
-    pub github_label: String,
-    pub sponsor_label: String,
 }
 
 impl Default for UiConfig {
@@ -119,13 +115,11 @@ impl Default for UiConfig {
                 description_label: "Description".to_string(),
                 submit_label: "Create".to_string(),
                 open_hint: "Press Ctrl-N to create a new memory".to_string(),
-                close_hint: "Tab: next field, Enter: submit, Esc: cancel".to_string(),
+                close_hint: "Tab: cycle fields, Enter: submit, Esc: back to tab focus".to_string(),
             },
             settings: SettingsOverlayText {
                 title: "Settings".to_string(),
-                theme_label: "Theme".to_string(),
-                theme_action_key: "t".to_string(),
-                close_hint: "Press Esc or S to close".to_string(),
+                close_hint: "Esc: close".to_string(),
             },
             help: HelpOverlayText {
                 title: "Help".to_string(),
@@ -135,12 +129,9 @@ impl Default for UiConfig {
                     "Esc: back / clear / close".to_string(),
                     "↑/↓ or j/k: move selection".to_string(),
                     "Enter or →: open/focus detail".to_string(),
-                    "o/c: open primary / secondary context links".to_string(),
                     "C: toggle chat panel".to_string(),
-                    "t: cycle theme".to_string(),
                     "?: toggle help".to_string(),
                     "q: quit".to_string(),
-                    "g/s: open GitHub / Sponsor".to_string(),
                 ],
                 close_hint: "Press any key to close".to_string(),
             },
@@ -148,8 +139,6 @@ impl Default for UiConfig {
                 commands_label: "Commands".to_string(),
                 tabs_label: "tabs".to_string(),
                 quit_label: "quit".to_string(),
-                github_label: "GitHub".to_string(),
-                sponsor_label: "Sponsor".to_string(),
             },
         }
     }
@@ -200,6 +189,7 @@ pub enum Focus {
     List,
     Tabs,
     Inspector,
+    Form,
     /// In-TUI chat panel (only when chat is open)
     Chat,
 }
@@ -217,6 +207,7 @@ impl Focus {
                     Focus::Tabs
                 }
             }
+            Focus::Form => Focus::Tabs,
             Focus::Chat => Focus::Tabs,
             Focus::Tabs => Focus::Search,
         }
@@ -228,6 +219,7 @@ impl Focus {
             Focus::Search => Focus::Tabs,
             Focus::List => Focus::Search,
             Focus::Inspector => Focus::List,
+            Focus::Form => Focus::Tabs,
             Focus::Chat => Focus::Inspector,
             Focus::Tabs => {
                 if chat_open {
